rust bindings: Update the build to link statically to librunecoral
TODO: Fix the build.rs for various build targets
