+++ rust/src/rune_coral.rs
@@ -1,64 +1,44 @@
 use crate::{ffi, Error, InferenceContext, TensorDescriptor};
 use std::{
-    ffi::{CString, OsStr},
+    ffi::{CString},
     mem::MaybeUninit,
-    ptr::{self, NonNull},
-    sync::Arc,
+    ptr::{self, NonNull}
 };
 
-/// A safe wrapper around `librunecoral` which has been dynamically loaded at
-/// runtime.
-pub struct RuneCoral {
-    inner: Arc<ffi::RuneCoral>,
-}
+pub fn create_inference_context(
+    mimetype: &str,
+    model: &[u8],
+    inputs: &[TensorDescriptor<'_>],
+    outputs: &[TensorDescriptor<'_>],
+) -> Result<InferenceContext, Error> {
+    let mimetype = CString::new(mimetype)?;
+    let mut inference_context = MaybeUninit::uninit();
 
-impl RuneCoral {
-    pub fn load(path: impl AsRef<OsStr>) -> Result<Self, libloading::Error> {
-        unsafe {
-            let inner = ffi::RuneCoral::new(path)?;
-            Ok(RuneCoral {
-                inner: Arc::new(inner),
-            })
-        }
-    }
+    let inputs = dummy_tensors(inputs);
+    let outputs = dummy_tensors(outputs);
 
-    pub fn create_inference_context(
-        &self,
-        mimetype: &str,
-        model: &[u8],
-        inputs: &[TensorDescriptor<'_>],
-        outputs: &[TensorDescriptor<'_>],
-    ) -> Result<InferenceContext, Error> {
-        let mimetype = CString::new(mimetype)?;
-        let mut inference_context = MaybeUninit::uninit();
+    // Safety: We've ensured our inputs are sane by construction (i.e. Rust
+    // doesn't let you create a null slice and all enums are exhaustive)
+    // and our `inputs` and `outputs` tensor vector can't outlive the
+    // `inputs` and `outputs` function arguments.
+    unsafe {
+        let ret = ffi::create_inference_context(
+            mimetype.as_ptr(),
+            model.as_ptr().cast(),
+            model.len() as ffi::size_t,
+            inputs.as_ptr(),
+            inputs.len() as ffi::size_t,
+            outputs.as_ptr(),
+            outputs.len() as ffi::size_t,
+            inference_context.as_mut_ptr(),
+        );
+        check_load_result(ret)?;
 
-        let inputs = dummy_tensors(inputs);
-        let outputs = dummy_tensors(outputs);
+        let inference_context = inference_context.assume_init();
 
-        // Safety: We've ensured our inputs are sane by construction (i.e. Rust
-        // doesn't let you create a null slice and all enums are exhaustive)
-        // and our `inputs` and `outputs` tensor vector can't outlive the
-        // `inputs` and `outputs` function arguments.
-        unsafe {
-            let ret = self.inner.create_inference_context(
-                mimetype.as_ptr(),
-                model.as_ptr().cast(),
-                model.len() as ffi::size_t,
-                inputs.as_ptr(),
-                inputs.len() as ffi::size_t,
-                outputs.as_ptr(),
-                outputs.len() as ffi::size_t,
-                inference_context.as_mut_ptr(),
-            );
-            check_load_result(ret)?;
-
-            let inference_context = inference_context.assume_init();
-
-            Ok(InferenceContext::new(
-                NonNull::new(inference_context).expect("Should be initialized"),
-                Arc::clone(&self.inner),
-            ))
-        }
+        Ok(InferenceContext::new(
+            NonNull::new(inference_context).expect("Should be initialized")
+        ))
     }
 }
 
@@ -103,14 +83,4 @@ fn dummy_tensors(inputs: &[TensorDescriptor<'_>]) -> Vec<ffi::RuneCoralTensor> {
     }
 
     tensors
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn library_wrapper_is_send_and_sync() {
-        static_assertions::assert_impl_all!(RuneCoral: Send, Sync);
-    }
 }
